refactor: rename delete-all-cart-items file to follow the action file name pattern
